refactor: booking state
