+++ components/BookingProgress.js
@@ -6,7 +6,7 @@ import Progress from './Progress';
 const BookingProgress = () => {
   const { steps } = useBookingState();
   const router = useRouter();
-  const currentStep = steps.find((step) => step.path === router.pathname);
+  const currentStep = steps.find((step) => router.pathname.includes(step));
   const currentStepIndex = steps.indexOf(currentStep);
   return <Progress progress={currentStepIndex + 1} total={steps.length} />;
 };
