+++ components/ListingPrice.js
@@ -3,13 +3,13 @@ import PropTypes from 'prop-types';
 import { __, priceFormat } from '../utils/helpers';
 import PlusIcon from './icons/PlusIcon';
 
-const ListingPrice = ({ fiatPrice, tokenPrice, utilityFiat, useToken }) => {
+const ListingPrice = ({ rentalFiat, rentalToken, utilityFiat, useTokens }) => {
   return (
     <div>
       <div className="flex items-center justify-center">
         <div className="flex-1 flex flex-col items-center">
           <p className="text-2xl leading-9">
-            {priceFormat(useToken ? tokenPrice : fiatPrice)}
+            {priceFormat(useTokens ? rentalToken : rentalFiat)}
           </p>
           <p>{__('bookings_price_switch_accomodation')}</p>
         </div>
@@ -26,11 +26,11 @@ const ListingPrice = ({ fiatPrice, tokenPrice, utilityFiat, useToken }) => {
 };
 
 ListingPrice.propTypes = {
-  fiatPrice: PropTypes.shape({
+  rentalFiat: PropTypes.shape({
     val: PropTypes.number,
     cur: PropTypes.string,
   }),
-  tokenPrice: PropTypes.shape({
+  rentalToken: PropTypes.shape({
     val: PropTypes.number,
     cur: PropTypes.string,
   }),
@@ -38,6 +38,7 @@ ListingPrice.propTypes = {
     val: PropTypes.number,
     cur: PropTypes.string,
   }),
+  useTokens: PropTypes.bool,
 };
 
 export default ListingPrice;
