+++ locales/base.json
@@ -169,6 +169,7 @@
   "newsletter_success": "Thanks, we will be in touch soon!",
   "newsletter_body": "Keep in the loop, subscribe:",
   "newsletter_signup": "Subscribe",
+  "page_error_title": "Error",
   "post_delete": "delete",
   "post_delete_message": "Post deleted.",
   "post_replies": "replies",
