+++ pages/bookings/[slug]/confirmation.js
