+++ pages/bookings/create/dates.js
@@ -8,13 +8,15 @@ import BookingGuests from '../../../components/BookingGuests';
 import BookingProgress from '../../../components/BookingProgress';
 import CurrencySwitch from '../../../components/CurrencySwitch';
 import Layout from '../../../components/Layout';
+import PageError from '../../../components/PageError';
 
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
+import PropTypes from 'prop-types';
 
-import blockchainConfig from '../../../config_blockchain.js';
 import { useAuth } from '../../../contexts/auth';
 import { useBookingActions, useBookingState } from '../../../contexts/booking';
+import api from '../../../utils/api';
 import { CURRENCIES, DEFAULT_CURRENCY } from '../../../utils/const';
 import { __ } from '../../../utils/helpers';
 
@@ -25,63 +27,61 @@ const defaultStart = dayjs()
   .set('hours', 18)
   .set('seconds', 0)
   .set('minutes', 0)
-  .toDate();
+  .format('YYYY-MM-DD');
+
 const defaultEnd = dayjs()
   .add(6, 'days')
   .set('hours', 11)
   .set('seconds', 0)
   .set('minutes', 0)
-  .toDate();
+  .format('YYYY-MM-DD');
 
-const DatesSelector = () => {
+const DatesSelector = ({ error, settings }) => {
   const router = useRouter();
   const { user } = useAuth();
   const isMember = user?.roles.includes('member');
-
-  const { steps, settings } = useBookingState();
-  const currentStep = steps.find((step) => step.path === router.pathname).data;
   const {
-    startDate: savedStartDate,
-    endDate: savedEndDate,
-    guests: savedGuests,
-    savedCurrency,
-  } = currentStep || {};
-  const { saveStepData, goToNextStep } = useBookingActions();
-  const [startDate, setStartDate] = useState(savedStartDate || defaultStart);
-  const [endDate, setEndDate] = useState(savedEndDate || defaultEnd);
-  const totalNights =
-    Math.abs(Math.ceil(dayjs(startDate).diff(endDate, 'days'))) + 1;
-
-  const [adults, setAdults] = useState(savedGuests?.adults || 1);
-  const [kids, setKids] = useState(savedGuests?.kids || 0);
-  const [infants, setInfants] = useState(savedGuests?.infants || 0);
-  const [pets, setPets] = useState(savedGuests?.pets || 0);
-
-  const [selectedCurrency, selectCurrency] = useState(
+    data: { dates: savedDates },
+  } = useBookingState();
+  const {
+    start: savedStartDate,
+    end: savedEndDate,
+    adults: savedAdults,
+    kids: savedKids,
+    infants: savedInfants,
+    pets: savedPets,
+    currency: savedCurrency,
+  } = savedDates || {};
+  const { saveStepData } = useBookingActions();
+  const [start, setStartDate] = useState(savedStartDate || defaultStart);
+  const [end, setEndDate] = useState(savedEndDate || defaultEnd);
+  const [adults, setAdults] = useState(savedAdults || 1);
+  const [kids, setKids] = useState(savedKids || 0);
+  const [infants, setInfants] = useState(savedInfants || 0);
+  const [pets, setPets] = useState(savedPets || 0);
+  const [currency, selectCurrency] = useState(
     savedCurrency || DEFAULT_CURRENCY,
   );
+  const totalNights = Math.abs(Math.ceil(dayjs(start).diff(end, 'days'))) + 1;
 
   const handleNext = () => {
-    saveStepData({
-      startDate,
-      endDate,
+    const data = {
+      start,
+      end,
       totalNights,
-      guests: {
-        adults,
-        kids,
-        infants,
-        pets,
-        totalGuests: adults,
-      },
-      useToken:
-        selectedCurrency === blockchainConfig.BLOCKCHAIN_DAO_TOKEN.symbol,
-      savedCurrency: selectedCurrency,
-    });
-    goToNextStep();
+      adults,
+      kids,
+      infants,
+      pets,
+      currency,
+    };
+    saveStepData(data);
+    const urlParams = new URLSearchParams(data);
+    router.push(`/bookings/create/accomodation?${urlParams}`);
   };
 
-  if (!settings.conditions) {
-    return null;
+  if (error) {
+    return <PageError error={error} />;
   }
 
   return (
@@ -95,9 +95,9 @@ const DatesSelector = () => {
         <BookingProgress />
         <div className="mt-16 flex flex-col gap-16">
           <BookingDates
-            conditions={settings.conditions}
-            startDate={startDate}
-            endDate={endDate}
+            conditions={settings?.conditions}
+            startDate={start}
+            endDate={end}
             setStartDate={setStartDate}
             setEndDate={setEndDate}
             isMember={isMember}
@@ -118,7 +118,7 @@ const DatesSelector = () => {
               <span>{__('bookings_dates_step_payment_title')}</span>
             </h2>
             <CurrencySwitch
-              selectedCurrency={selectedCurrency}
+              selectedCurrency={currency}
               onSelect={selectCurrency}
               currencies={CURRENCIES}
             />
@@ -130,6 +130,42 @@ const DatesSelector = () => {
       </div>
     </Layout>
   );
+};
+
+export const getStaticProps = async () => {
+  try {
+    const {
+      data: { results },
+    } = await api.get('/bookings/settings');
+    return {
+      props: {
+        settings: results,
+      },
+    };
+  } catch (err) {
+    return {
+      props: {
+        error: err.message,
+        questions: null,
+      },
+    };
+  }
+};
+
+DatesSelector.propTypes = {
+  error: PropTypes.string,
+  settings: PropTypes.shape({
+    conditions: PropTypes.shape({
+      member: PropTypes.shape({
+        maxBookingHorizon: PropTypes.number,
+        maxDuration: PropTypes.number,
+      }),
+      guest: PropTypes.shape({
+        maxBookingHorizon: PropTypes.number,
+        maxDuration: PropTypes.number,
+      }),
+    }),
+  }),
 };
 
 export default DatesSelector;
