+++ pages/bookings/create/index.js
@@ -5,16 +5,14 @@ import React from 'react';
 
 import PageNotAllowed from '../../401';
 import { useAuth } from '../../../contexts/auth';
-import { BOOKING_PATHS } from '../../../utils/const';
 
 const NewBooking = () => {
   const router = useRouter();
   const { isAuthenticated } = useAuth();
-  const [firstStepUrl] = BOOKING_PATHS;
 
   useEffect(() => {
     if (isAuthenticated) {
-      router.push(firstStepUrl);
+      router.push('/bookings/create/dates');
     }
   }, [isAuthenticated]);
