+++ pages/bookings/index.js
@@ -1,11 +1,12 @@
 import Head from 'next/head';
+
+import Bookings from '../../components/Bookings';
 import Layout from '../../components/Layout';
+import Tabs from '../../components/Tabs';
 
 import PageNotFound from '../404';
 import { useAuth } from '../../contexts/auth';
 import { __ } from '../../utils/helpers';
-import Tabs from '../../components/Tabs';
-import Bookings from '../../components/Bookings';
 
 const BookingsDirectory = () => {
   const { user } = useAuth();
@@ -16,16 +17,16 @@ const BookingsDirectory = () => {
         createdBy: user._id,
         status: ['pending', 'confirmed', 'checkedIn'],
         end: {
-          $gt: new Date()
-        }
-      }
+          $gt: new Date(),
+        },
+      },
     },
     pastBookings: user && {
       where: {
         createdBy: user._id,
-        end: { $lt: new Date() }
-      }
-    }
+        end: { $lt: new Date() },
+      },
+    },
   };
 
   if (!user) {
@@ -39,18 +40,18 @@ const BookingsDirectory = () => {
       </Head>
       <div className="main-content intro fullwidth">
         <Tabs
-          tabs={ [
+          tabs={[
             {
               title: __('bookings_title'),
               value: 'my-bookings',
-              content: <Bookings filter={ filters.myBookings } />
+              content: <Bookings filter={filters.myBookings} />,
             },
             {
               title: __('past_bookings_title'),
               value: 'past-bookings',
-              content: <Bookings filter={ filters.pastBookings } />
+              content: <Bookings filter={filters.pastBookings} />,
             },
-          ] }
+          ]}
         />
       </div>
     </Layout>
