+++ utils/const.js
@@ -3,13 +3,13 @@
 */
 import { BLOCKCHAIN_DAO_TOKEN } from '../config_blockchain.js';
 
-export const BOOKING_PATHS = [
-  '/bookings/new/dates',
-  '/bookings/new/accomodation',
-  '/bookings/new/questionnaire',
-  '/bookings/new/summary',
-  '/bookings/new/checkout',
-  '/bookings/new/confirmation',
+export const BOOKING_STEPS = [
+  '/bookings/create/dates',
+  '/bookings/create/accomodation',
+  '/bookings/questionnaire',
+  '/bookings/summary',
+  '/bookings/checkout',
+  '/bookings/confirmation',
 ];
 
 export const CURRENCIES = ['EUR', BLOCKCHAIN_DAO_TOKEN.symbol];
